added some more instruction and made the pixel library to draw what inside the display array by converting each bit to a pixel of a std image/png and draw the png using pixel sprite

+++ internal/chip8/registers.go
@@ -6,7 +6,7 @@ type Registers struct {
 	Delay byte
 	Sound byte
 
-	I byte
+	I uint16
 	Sp byte
 	Pc uint16
 }
